Clean up eval scenarios and stale docs
- Remove 7 unused/redundant scenarios (compaction_stress, phase2_compaction,
  basic_2step_stateful_tre, compaction_stress_stateful,
  phase2_compaction_stateful, inventory_audit, supplier_deep_dive)
- Keep 22 scenarios: 18 copy-paste + 4 compaction chain
- Extract relevance_detection_stateful into _stateful_relevance.py
- Delete _stateful_compaction.py
- Update test counts: 800 (README, CONTRIBUTING)
- Update scenario counts: 22 (README, EVAL_GUIDE)
- Remove roadmap section from README (items shipped)
- Mark ADR-006 as implemented
- Update test_eval_budget.py to use compaction_chain_p1
