+++ CONTRIBUTING.md
@@ -16,7 +16,7 @@ pip install -e ".[dev]"
 Unit tests are fully deterministic — no LLM backend required.
 
 ```bash
-# Full suite (562 tests)
+# Full suite (800 tests)
 python -m pytest tests/unit/ -v --tb=short
 
 # With coverage
