+++ README.md
@@ -141,7 +141,7 @@ python -m pytest tests/ --cov=forge --cov-report=term-missing
 
 ## Eval Harness
 
-31 scenarios (18 in batch eval results) measuring how reliably a model + backend combo navigates multi-step tool-calling workflows. See [Eval Guide](docs/EVAL_GUIDE.md) for full CLI reference.
+22 scenarios measuring how reliably a model + backend combo navigates multi-step tool-calling workflows. See [Eval Guide](docs/EVAL_GUIDE.md) for full CLI reference.
 
 ```bash
 # Ollama
@@ -197,17 +197,9 @@ src/forge/
     handler.py         # Request handler — bridge between HTTP and run_inference
     convert.py         # OpenAI messages ↔ forge Messages conversion
 tests/
-  unit/                # 655 deterministic tests — no LLM backend required
+  unit/                # 800 deterministic tests — no LLM backend required
   eval/                # Eval harness — model qualification against real backends
 ```
-
-## Roadmap
-
-1. **Multi-model routing** — Model pool for managing N backends simultaneously. See [`docs/decisions/MULTI_MODEL_ROUTING.md`](docs/decisions/MULTI_MODEL_ROUTING.md).
-2. **Tool prerequisites** — Conditional tool dependencies. See [`docs/decisions/006-tool-prerequisites.md`](docs/decisions/006-tool-prerequisites.md).
-3. **Context window self-awareness** — Inject remaining context budget so the model knows compaction is approaching.
-4. **Compaction tiers** — Consumer-configurable per-phase compaction thresholds.
-5. **Proxy server enhancements** — Request queuing for concurrent clients, client disconnect cancellation, text response intent (ADR-013).
 
 ## Documentation
