+++ tests/eval/scenarios/__init__.py
@@ -1,7 +1,7 @@
 """Eval scenario definitions — re-exports from submodules."""
 
 from ._base import EvalScenario, _check
-from ._compaction import phase2_compaction, relevance_detection
+from ._compaction import relevance_detection
 from ._model_quality import (
     argument_fidelity,
     conditional_routing,
@@ -9,19 +9,12 @@ from ._model_quality import (
     sequential_reasoning,
     tool_selection,
 )
-from ._plumbing import basic_2step, compaction_stress, error_recovery, sequential_3step
+from ._plumbing import basic_2step, error_recovery, sequential_3step
 from ._compaction_chain import (
     compaction_chain_baseline,
     compaction_chain_p1,
     compaction_chain_p2,
     compaction_chain_p3,
-)
-from ._stateful_compaction import (
-    compaction_stress_stateful,
-    inventory_audit,
-    phase2_compaction_stateful,
-    relevance_detection_stateful,
-    supplier_deep_dive,
 )
 from ._stateful_model_quality import (
     argument_fidelity_stateful,
@@ -32,26 +25,24 @@ from ._stateful_model_quality import (
 )
 from ._stateful_plumbing import (
     basic_2step_stateful,
-    basic_2step_stateful_tre,
     error_recovery_stateful,
     sequential_3step_stateful,
 )
+from ._stateful_relevance import relevance_detection_stateful
 
 ALL_SCENARIOS: list[EvalScenario] = [
+    # Lambda (non-stateful)
     basic_2step,
     sequential_3step,
-    compaction_stress,
     error_recovery,
     tool_selection,
     argument_fidelity,
     sequential_reasoning,
     conditional_routing,
     data_gap_recovery,
-    phase2_compaction,
     relevance_detection,
-    # Stateful scenarios
+    # Stateful
     basic_2step_stateful,
-    basic_2step_stateful_tre,
     sequential_3step_stateful,
     error_recovery_stateful,
     tool_selection_stateful,
@@ -59,11 +50,8 @@ ALL_SCENARIOS: list[EvalScenario] = [
     sequential_reasoning_stateful,
     conditional_routing_stateful,
     data_gap_recovery_stateful,
-    compaction_stress_stateful,
-    phase2_compaction_stateful,
-    inventory_audit,
-    supplier_deep_dive,
     relevance_detection_stateful,
+    # Compaction chain
     compaction_chain_baseline,
     compaction_chain_p1,
     compaction_chain_p2,
@@ -76,17 +64,14 @@ __all__ = [
     "ALL_SCENARIOS",
     "basic_2step",
     "sequential_3step",
-    "compaction_stress",
     "error_recovery",
     "tool_selection",
     "argument_fidelity",
     "sequential_reasoning",
     "conditional_routing",
     "data_gap_recovery",
-    "phase2_compaction",
     "relevance_detection",
     "basic_2step_stateful",
-    "basic_2step_stateful_tre",
     "sequential_3step_stateful",
     "error_recovery_stateful",
     "tool_selection_stateful",
@@ -94,10 +79,6 @@ __all__ = [
     "sequential_reasoning_stateful",
     "conditional_routing_stateful",
     "data_gap_recovery_stateful",
-    "compaction_stress_stateful",
-    "phase2_compaction_stateful",
-    "inventory_audit",
-    "supplier_deep_dive",
     "relevance_detection_stateful",
     "compaction_chain_baseline",
     "compaction_chain_p1",
